requirements.txt
python-dotenv
google-generativeai
langchain
PyPDF2
faiss-cpu
langchain_community
streamlit
openai
-e .